Move MemoryExtensions.TryGetString to MemoryMarshal and remove TryGetArray
